Improve error reporting by not stopping on first reported error
- Multiple error reporting with error cascade avoidance
- Add type checks for instruction encoding values (e.g., that the
  immediate field is an int, not a string)
- In general, more robust error handling.

Also some language improvements:

- Allow !include to take an expression as the filename
  (so you could do !include ("foo" + ".asm"))
- Ditto for !binary

Error reporting:

Add ability to report multiple errors on a compile rather than
erroring out on the first problem.  This is quite handy when
refactoring code: for example if you rename a variable, it's
good to be able to tell where it was used, and not just the
first use.

This is implemented in such a way that "error cascades" are
reasonably well avoided.  An error cascade is a situation
in which some expression or other code is dependent on a value
that itself contained an error.  For example:

!let idx = xx          ; undeclared variable xx
!let arr = [0, 1, 2]
lda #arr[idx]          ; skip error reporting for 'idx' resolve

The last line might emit spurious, non-sensible errors (like wrong type
'idx' for the array access).  So errors are tracked with expression
values and error reporting suppressed in situations where operands
have errors in them.

This change also adds a number of new test cases.

Test runner changes:

- Require that error tests contain the same number of expected
  and actual errors.