Better *=addr handling, add warnings support
Earlier the assembler and the disassembler assumed
the code always starts at address $801.  This only works
for C64.

Add handling to default to this starting address if it's
not been specified with *=addr.

If it's never specified in the source, then default to
a good default ($801 for C64) but emit a warning about it.

Since we're now emitting warnings, add test support for checking
correctly emitted warnings.