Handle 'file not found' in parse cache and guardedReadFileSync()
Because guardedReadFileSync pretended to succeed when failing
to load a file, errors such as missing file for !include
were sometimes not reported as they'd get masked out by
the caches returning dummy values instead of undefined.

+ attempt to add a test case for a weird scoping bug while
trying to compile bintris with c64jasm.  This was not it,
but might as well leave the testcase in the test suite.