Fix tests to use debuginfo'd disassembly
Some tests were crashing in disassembly as the disassembler
tries to decode some byte sequences and reads past the input
buffer.  These bugs didn't trigger before illegal opcodes
were added.  This should be fixed in the disassembler
but for now I'm fixing the tests to be more sane first.

The change passes debug info from assembly to disassembly,
so that raw byte arrays can be decoded as bytes without
attempts to disassemble them as opcodes.

for #79